% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_indicator_cols}
\alias{add_indicator_cols}
\title{Add columns indicating whether the event took plays in y days before given date}
\usage{
add_indicator_cols(x, event_date_col, end_date_col, intervals)
}
\arguments{
\item{x}{data frame}

\item{event_date_col}{Character, name of the column in which event dates are stored}

\item{end_date_col}{Character, name of the column in which end_date is stored}

\item{intervals}{Numeric vector of day intervals. Events that occur in the span of y days for
each y in interval will be calculated}
}
\value{
data frame
}
\description{
Add columns indicating whether the event took plays in y days before given date
}
