% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_events.R
\name{count_events}
\alias{count_events}
\title{Count events for all cases in the database}
\usage{
count_events(x, case_id_col = "spr_Id",
  event_date_col = "evl_EventDate", end_date_col = "Data0",
  event_codes_cols = c("ert_Code", "ser_Code", "ier_Code"),
  intervals = c(30, 90, 180, 360), groupings = list(1, 2, 3, 1:2, 1:3,
  c(1, 3)))
}
\arguments{
\item{x}{Data frame}

\item{case_id_col}{Character, name of the column in which case IDs are stored}

\item{event_date_col}{Character, name of the column where event dates are stored}

\item{end_date_col}{Character, name of the column in which end_date is stored}

\item{event_codes_cols}{Characters vector of names of column where action codes are stored}

\item{intervals}{Numeric vector of day intervals. Events that occur in the span of y days for
each y in interval will be calculated}

\item{groupings}{list of vectors of combinations of positions in event_codes_cols by which
grouping will be performed}
}
\value{
list of data frames
}
\description{
Count events for all cases in the database
}
