% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_events.R
\name{count_events_single}
\alias{count_events_single}
\title{Count events for a single case}
\usage{
count_events_single(x, case_id, case_id_col, event_date_col, end_date_col,
  event_codes_cols, intervals, groupings)
}
\arguments{
\item{x}{Data frame}

\item{case_id}{ID of the case for which events will be counted}

\item{case_id_col}{Character, name of the column in which case IDs are stored}

\item{event_date_col}{Character, name of the column where event dates are stored}

\item{end_date_col}{Character, name of the column in which end_date is stored}

\item{event_codes_cols}{Characters vector of names of column where action codes are stored}

\item{intervals}{Numeric vector of day intervals. Events that occur in the span of y days for
each y in interval will be calculated}

\item{groupings}{list of vectors of combinations of positions in event_codes_cols by which
grouping will be performed}
}
\value{
data frame
}
\description{
Count events for a single case
}
