% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{explore_interactions_rf}
\alias{explore_interactions_rf}
\title{Find interacting variables by fitting random forest}
\usage{
explore_interactions_rf(x, y)
}
\arguments{
\item{x}{explanatory variables, all factors}

\item{y}{response variable}
}
\value{
data frame
}
\description{
Find interacting variables by fitting random forest
}
