% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fill_na}
\alias{fill_na}
\title{Fill NAs that were casued by merging dfs of differing ncols}
\usage{
fill_na(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
Fill NAs that were casued by merging dfs of differing ncols
}
