% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_counts}
\alias{find_counts}
\title{Calculate numbers of event in time intervals}
\usage{
find_counts(to_count, grouping_variables, case_id_col)
}
\arguments{
\item{to_count}{data frame}

\item{grouping_variables}{Character, names of variables to group by}

\item{case_id_col}{Character, name of the column in which case IDs are stored}
}
\value{
data frame
}
\description{
Calculate numbers of event in time intervals
}
