% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_first_last}
\alias{find_first_last}
\title{Find date of first and last occurence of the event}
\usage{
find_first_last(to_count, grouping_variables, case_id_col, event_date_col)
}
\arguments{
\item{to_count}{data frame}

\item{grouping_variables}{Character, names of columns to group by}

\item{case_id_col}{Character, name of the column in which case IDs are stored}

\item{event_date_col}{Character, name of the column that contains dates of events}
}
\value{
data frame
}
\description{
Find date of first and last occurence of the event
}
