% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{get_all_interactions}
\alias{get_all_interactions}
\title{Get interactions terms for a set of categorical predictors}
\usage{
get_all_interactions(x, exclude = NULL)
}
\arguments{
\item{x}{data frame}

\item{exclude}{character or factor columns to exclude}
}
\value{
model matrix
}
\description{
Get interactions terms for a set of categorical predictors
}
