% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{get_factors_summary}
\alias{get_factors_summary}
\title{Get summary of classes and unique values in factors or character variables}
\usage{
get_factors_summary(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
Get summary of classes and unique values in factors or character variables
}
