% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{lump_wisely}
\alias{lump_wisely}
\title{Lump together rare levels only if number of levels changed}
\usage{
lump_wisely(fct, min_prop)
}
\arguments{
\item{fct}{factor of character vector}

\item{min_prop}{minimum proportion to leave level unchanged}
}
\value{
factor vector
}
\description{
Lump together rare levels only if number of levels changed
}
