% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_wide}
\alias{make_wide}
\title{Transform data into wide form}
\usage{
make_wide(src_df, grouping_variables, case_id_col)
}
\arguments{
\item{src_df}{data frame}

\item{grouping_variables}{Character, names of grouping variable}

\item{case_id_col}{Character, name of the column in which case IDs are stored}
}
\value{
data frame
}
\description{
Transform data into wide form
}
