% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{merge_levels_rare}
\alias{merge_levels_rare}
\title{Merge rare levels of all factors in data frame}
\usage{
merge_levels_rare(x, exclude = NULL, min_n = NULL, min_prop = NULL)
}
\arguments{
\item{x}{data frame}

\item{exclude}{character or factor columns to exclude}

\item{min_n}{Levels with less than n observations will be merged}

\item{min_prop}{Levels with observations that make up less than 100*min_prop% 
will be merged.}
}
\value{
factor vector
}
\description{
Merge rare levels of all factors in data frame
}
