% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{merge_levels_response}
\alias{merge_levels_response}
\title{Merge levels of factors based on relationship to response variable}
\usage{
merge_levels_response(fct, response)
}
\arguments{
\item{fct}{Factor to merge}

\item{response}{Response variable, logical vector}
}
\value{
factor vector
}
\description{
Merge levels of factors based on relationship to response variable
}
