% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{replace_na}
\alias{replace_na}
\title{Replace missing values with a special token}
\usage{
replace_na(x, replacement = "brak_inf")
}
\arguments{
\item{x}{Factor or character vector}

\item{replacement}{String which will replace missing values}
}
\value{
data frame
}
\description{
Replace missing values with a special token
}
