% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{single_grouping}
\alias{single_grouping}
\title{Find counts and first/last event date for a single set of grouping variables}
\usage{
single_grouping(x, grouping_variables, case_id_col, event_date_col,
  end_date_col, intervals)
}
\arguments{
\item{x}{Data frame}

\item{grouping_variables}{Character, names of columns to group by}

\item{case_id_col}{Character, name of the column in which case IDs are stored}

\item{event_date_col}{Character, name of the column where event dates are stored}

\item{end_date_col}{Character, name of the column in which end_date is stored}

\item{intervals}{Numeric vector of day intervals. Events that occur in the span of y days for
each y in interval will be calculated}
}
\value{
data frame
}
\description{
Find counts and first/last event date for a single set of grouping variables
}
